requirements.txt
asgiref==3.2.3
asn1crypto==0.24.0
attrs==17.4.0
Automat==0.6.0
beautifulsoup4==4.8.2
blinker==1.4
bs4==0.0.1
certifi==2018.1.18
chardet==3.0.4
click==6.7
cloud-init==19.1
colorama==0.3.7
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.1.4
cycler==0.10.0
decorator==4.4.1
distro-info===0.18ubuntu0.18.04.1
Django==3.0.2
httplib2==0.9.2
hyperlink==17.3.1
idna==2.6
incremental==16.10.1
Jinja2==2.10
jsonpatch==1.16
jsonpointer==1.10
jsonschema==2.6.0
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.1.0
language-selector==0.1
MarkupSafe==1.0
matplotlib==3.1.2
netifaces==0.10.4
networkx==2.4
numpy==1.18.1
oauthlib==2.0.6
PAM==0.4.2
Pillow==7.0.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycrypto==2.6.1
pygobject==3.26.1
PyJWT==1.5.3
pyOpenSSL==17.5.0
pyparsing==2.4.6
PyQt5==5.14.1
PyQt5-sip==12.7.0
pyserial==3.4
python-apt==1.6.4
python-dateutil==2.8.1
python-debian==0.1.32
pytz==2018.3
pyxdg==0.25
PyYAML==3.12
requests==2.18.4
requests-unixsocket==0.1.5
SecretStorage==2.3.1
service-identity==16.0.0
six==1.11.0
soupsieve==1.9.5
sqlparse==0.2.4
ssh-import-id==5.7
systemd-python==234
tornado==6.0.3
Twisted==17.9.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.22
zope.interface==4.3.2
